fixed directories
