+++ frontend/src/components/UserDash/applicantInfoHelpers.js
@@ -15,12 +15,6 @@ const defaultRole = {
   }
 };
 
-const defaultMandatedHours = {
-  get: function(target, name) {
-    return target.hasOwnProperty(name) ? target[name] : 'Not Mandated';
-  }
-};
-
 export const statuses = new Proxy(
   {
     has_volunteered: 'Has Volunteered',
@@ -34,14 +28,6 @@ export const roles = new Proxy(
     admin: 'Administrator',
     manager: 'Manager',
     volunteer: 'Volunteer'
-  },
-  defaultRole
-);
-
-export const mandated = new Proxy(
-  {
-    is_mandated: 'Mandated Hours',
-    not_mandated: 'Not Mandated'
   },
   defaultRole
 );
